+++ MovingObj.hpp
@@ -8,14 +8,14 @@ public :
     ~MovingObj();
 
 protected :
-    int a_x = 0, a_y = 0;
-    int v_Obj = 0;
+    double a_x = 0, a_y = 0;
+    double v_Obj = 0;
 public :
     int damage = 0;
     bool exist = 0;
     int c_x, c_y;
 public :
-    void Change( int ax, int ay );
+    void Change( double ax, double ay );
     void Located();
 //    void Run();
 };
